Add latch to de-dupe events from multiple sidecars.
If we allow all Sidecars in the same cluster to talk to
the same Superside, we'll end up with lots of duped events,
but we only really want one event stream. This is a really
simple way to solve it. Throughput should not suffer much
and it takes all the work off of Sidecar.